requirements.txt
streamlit
opencv-python-headless
numpy
plotly
Pillow